pyproject.toml
[project]
name = "mycelium"
version = "0.1.0"
description = "Distributed OS for collective consciousness"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.18",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "typer>=0.12",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "json-repair>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.0",
]
local-ml = [
    "fastembed>=0.4",
]
mcp = [
    "fastmcp>=2.0",
]
render = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
]
full = [
    "mycelium[mcp,local-ml,render]",
]

[project.scripts]
mycelium = "mycelium.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mycelium"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires running Neo4j (deselect: -m 'not integration')",
    "semantic: requires local BGE-M3 model / local-ml extra (deselect: -m 'not semantic')",
    "extraction: requires claude CLI (deselect: -m 'not extraction')",
    "benchmark: performance benchmarks (includes integration tests)",
]

[tool.ruff]
target-version = "py312"
src = [".", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["mycelium"]
mypy_path = "."

[[tool.mypy.overrides]]
module = ["structlog", "structlog.*", "neo4j", "neo4j.*", "fastembed", "fastembed.*", "fastmcp", "fastmcp.*", "fastapi", "fastapi.*", "uvicorn", "uvicorn.*"]
ignore_missing_imports = true

